Pause on Autosave - Quality of life improvement
When playing with friends, the autosave feature would often be a cause of frustration, because some players' computers would complete the save process before others, and the game would resume without them. Those players would then need to catch up. Under normal circumstances this would be okay, but when in a moment of stress, like during combat, this could cause some real issues for the players who were behind. This change allows the host to toggle a
new option that simply pauses the world when an autosave happens. Once complete, any user can resume the game as normal.